*: add composer json generation and CI/CD workflow
PHP libraries use a composer.json file to declare their namespaces and
their location. This commit adds a bazel script to automatically
generate this composer.json file, and a sanity-checking script to
validate that this file is up to date with the current source code. It
also includes a GitHub Workflow to run all these checks.